[refactor] 반복코드 제거
