+++ src/components/frame/MobileHeader.tsx
@@ -23,9 +23,11 @@ import { menuData } from '@/assets/navMenuData';
 export const MobileHeader = ({
   isHome,
   isScrolled,
+  mounted,
 }: {
   isHome: boolean;
   isScrolled: boolean;
+  mounted: boolean;
 }) => {
   const [open, setOpen] = useState(false);
 
@@ -70,7 +72,8 @@ export const MobileHeader = ({
   return (
     <div
       className={cn(
-        'md:hidden fixed top-0 left-0 right-0 flex items-center justify-between w-full h-14 text-center px-10 py-4 z-30 transition-all duration-300',
+        'md:hidden fixed top-0 left-0 right-0 flex items-center justify-between w-full h-14 text-center px-10 py-4 z-30 transition-all duration-500 ease-in-out',
+        mounted ? 'translate-y-0 opacity-100' : '-translate-y-10 opacity-0',
         getStyle('header')
       )}
     >
